+++ src/tasks/mod.rs
@@ -53,6 +53,7 @@ mod tmuxconf;
 mod vale;
 mod vim;
 mod vscode;
+mod vscodejson;
 mod windows;
 mod yq;
 mod zsh;
@@ -150,14 +151,15 @@ fn sequence() -> Vec<String> {
         googlecloudsdk::task().name,
         hyper::task().name, // deps: config
         macos::task().name,
-        psql::task().name,     // deps: config
-        ssh::task().name,      // deps: config
-        tmuxconf::task().name, // deps: config
-        tmux::task().name,     // deps: tmuxconf,brewbundle
-        vim::task().name,      // deps: config,pip; takes over the terminal
-        vscode::task().name,   // deps: config
-        zshrc::task().name,    // deps: config
-        zsh::task().name,      // deps: profile,inputrc,zshrc,brewbundle
+        psql::task().name,       // deps: config
+        ssh::task().name,        // deps: config
+        tmuxconf::task().name,   // deps: config
+        tmux::task().name,       // deps: tmuxconf,brewbundle
+        vim::task().name,        // deps: config,pip; takes over the terminal
+        vscodejson::task().name, // deps: config
+        vscode::task().name,     // deps: config,vscodejson
+        zshrc::task().name,      // deps: config
+        zsh::task().name,        // deps: profile,inputrc,zshrc,brewbundle
         windows::task().name,
     ]
 }
@@ -214,6 +216,7 @@ fn mapping() -> HashMap<String, Task> {
     map.insert(String::from("vale"), vale::task());
     map.insert(String::from("vim"), vim::task());
     map.insert(String::from("vscode"), vscode::task());
+    map.insert(String::from("vscodejson"), vscodejson::task());
     map.insert(String::from("windows"), windows::task());
     map.insert(String::from("yq"), yq::task());
     map.insert(String::from("zsh"), zsh::task());